refactor responses and requests
