+++ postrify-backend/src/main/java/com/postrify/postrifybackend/dto/MessageResponse.java
@@ -1,4 +1,4 @@
-package com.postrify.postrifybackend.model;
+package com.postrify.postrifybackend.dto;
 
 public class MessageResponse {
   private String message;
