+++ postrify-backend/src/main/java/com/postrify/postrifybackend/dto/SignUpRequest.java
@@ -1,4 +1,4 @@
-package com.postrify.postrifybackend.model;
+package com.postrify.postrifybackend.dto;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
